+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     from the controller) requires an array with the `user` key as configuration;
 * improved and fixed the `KcFinderComponent`;
 * improved validation rules. Removed some useless validation methods;
+* uses `league/commonmark` package instead of `gourmet/common-mark`;
 * added `MeCms\AuthTrait`. It provides some methods for classes that need to
     verify the data of the logged in user;
 * added `MeCms\Validation\BannerAndPhotoValidator` abstract class;